Adult card component split-out & cleanup
